+++ docs/changelog.md
@@ -7,3 +7,4 @@
 * `build-code-coverage-metrics` - covert PHPUnit clover.xml into Markdown and badges
 * `build-contracts` - convert PHPUnit testdox.txt into Markdown
 * `build-php-badges` - get PHP badges from shields.io
+* `bengi.json` config file for your project